requirements.txt
Django==1.7.1
Pillow==2.6.1
gunicorn==19.1.1
